Major changes: 1) Central difference changed from 4 to 2 (faster). 2) Hyperparameter optimization on. 3) Acq. function uncertainty (kappa=4). 4) New function to redistribute the images along the path after being optimize. Fixed some bugs in print. General cleaning.

+++ catlearn/optimize/plots.py
@@ -160,7 +160,7 @@ def get_plots_neb(images, selected=None, iter=None):
     ----------
     images : Atoms objects or trajectory file.
         Atoms objects or traj file (in ASE format) containing the path to plot.
-    selected: float
+    selected: int
         Next point to evaluate, chosen by the surrogate model.
     iter: integer
         Iteration number.
